refactor : 안쓰는 코드와 경로수정
