+++ js/state/state.js
@@ -1,26 +1,26 @@
 export const sharedState = {
   //수입 지출 요소
-  isIncome: true, // true면 수입, false면 지출
+  isIncome: true,
   selectedMethod: null,
   selectedCategory: null,
 
   // 필터링 상태
-  showIncome: true, // 수입 내역 표시 여부
-  showExpense: true, // 지출 내역 표시 여부
+  showIncome: true,
+  showExpense: true,
   totalIncome: 0,
   totalExpense: 0,
-  entryId: null, // 현재 수정 중인 항목의 ID
+  entryId: null,
 
-  entries: [], // 서버에서 가져온 항목들을 저장할 배열
+  entries: [],
 
   // 뷰 관련 상태
-  activeView: "ledger", // 현재 활성화된 뷰 (ledger, calendar, stats)
-  calendarInitialized: false, // 캘린더 뷰 초기화 여부
-  statsInitialized: false, // 통계 뷰 초기화 여부
+  activeView: "ledger",
+  calendarInitialized: false,
+  statsInitialized: false,
 };
 
 import { DateStore } from "./store.js";
 export const dateStore = new DateStore({
   year: new Date().getFullYear(),
-  month: new Date().getMonth() + 1, // 월은 0부터
+  month: new Date().getMonth() + 1,
 });
